% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods.R
\name{confint.ctmeddelta}
\alias{confint.ctmeddelta}
\title{Delta Method Confidence Intervals}
\usage{
\method{confint}{ctmeddelta}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Object of class \code{ctmeddelta}.}

\item{parm}{a specification of which parameters
are to be given confidence intervals,
either a vector of numbers or a vector of names.
If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of confidence intervals.
}
\description{
Delta Method Confidence Intervals
}
\examples{
data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Specific time-interval ----------------------------------------------------
delta <- DeltaMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m"
)
confint(delta)

# Range of time-intervals ---------------------------------------------------
delta <- DeltaMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m"
)
confint(delta)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
