% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods.R
\name{confint.ctmedmc}
\alias{confint.ctmedmc}
\title{Monte Carlo Method Confidence Intervals}
\usage{
\method{confint}{ctmedmc}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Object of class \code{ctmedmc}.}

\item{parm}{a specification of which parameters
are to be given confidence intervals,
either a vector of numbers or a vector of names.
If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of confidence intervals.
}
\description{
Monte Carlo Method Confidence Intervals
}
\examples{
data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Specific time-interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
confint(mc)

# Range of time-intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
confint(mc)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
