% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-delta-med.R
\name{DeltaMed}
\alias{DeltaMed}
\title{Delta Method Sampling Variance-Covariance Matrix
for the Total, Direct, and Indirect Effects of X on Y
Through M
Over a Specific Time-Interval
or a Range of Time-Intervals}
\usage{
DeltaMed(phi, vcov_phi_vec, delta_t, from, to, med, ncores = NULL)
}
\arguments{
\item{phi}{Numeric matrix.
The drift matrix (\eqn{\boldsymbol{\Phi}}).
\code{phi} should have row and column names
pertaining to the variables in the system.}

\item{vcov_phi_vec}{Numeric matrix.
The sampling variance-covariance matrix of
\eqn{\mathrm{vec} \left( \boldsymbol{\Phi} \right)}.}

\item{delta_t}{Vector of positive numbers.
Time interval
(\eqn{\Delta t}).}

\item{from}{Character string.
Name of the independent variable \eqn{X} in \code{phi}.}

\item{to}{Character string.
Name of the dependent variable \eqn{Y} in \code{phi}.}

\item{med}{Character vector.
Name/s of the mediator variable/s in \code{phi}.}

\item{ncores}{Positive integer.
Number of cores to use.
If \code{ncores = NULL},
use a single core.
Consider using multiple cores
when the length of \code{delta_t} is long.}
}
\value{
Returns an object
of class \code{ctmeddelta} which is a list with the following elements:
\describe{
\item{call}{Function call.}
\item{args}{Function arguments.}
\item{fun}{Function used (DeltaMed).}
\item{output}{A list with length of \code{length(delta_t)}.}
}
Each element in the \code{output} list has the following elements:
\describe{
\item{delta_t}{Time-interval.}
\item{jacobian}{Jacobian matrix.},
\item{est}{Estimated total, direct, and indirect effects.},
\item{vcov}{Sampling variance-covariance matrix of the
estimated total, direct, and indirect effects.}
}
}
\description{
This function computes the delta method
sampling variance-covariance matrix
for the total, direct, and indirect effects
of the independent variable \eqn{X}
on the dependent variable \eqn{Y}
through mediator variables \eqn{\mathbf{m}}
over a specific time-interval \eqn{\Delta t}
or a range of time-intervals
using the first-order stochastic differential equation model's
drift matrix \eqn{\boldsymbol{\Phi}}.
}
\details{
See \code{\link[=Total]{Total()}},
\code{\link[=Direct]{Direct()}}, and
\code{\link[=Indirect]{Indirect()}} for more details.
\subsection{Delta Method}{

Let \eqn{\boldsymbol{\theta}} be
\eqn{\mathrm{vec} \left( \boldsymbol{\Phi} \right)},
that is,
the elements of the \eqn{\boldsymbol{\Phi}} matrix
in vector form sorted column-wise.
Let \eqn{\hat{\boldsymbol{\theta}}} be
\eqn{\mathrm{vec} \left( \hat{\boldsymbol{\Phi}} \right)}.
By the multivariate central limit theory,
the function \eqn{\mathbf{g}}
using \eqn{\hat{\boldsymbol{\theta}}} as input
can be expressed as:

\deqn{
  	\sqrt{n}
  	\left(
  	\mathbf{g} \left( \hat{\boldsymbol{\theta}} \right)
  	-
  	\mathbf{g} \left( \boldsymbol{\theta} \right)
  	\right)
  	\xrightarrow[]{
  		\mathrm{D}
  	}
  	\mathcal{N}
  	\left(
  	0,
  	\mathbf{J}
  	\boldsymbol{\Gamma}
  	\mathbf{J}^{\prime}
  	\right)
  }

where \eqn{\mathbf{J}} is the matrix of first-order derivatives
of the function \eqn{\mathbf{g}}
with respect to the elements of \eqn{\boldsymbol{\theta}}
and
\eqn{\boldsymbol{\Gamma}}
is the asymptotic variance-covariance matrix of
\eqn{\hat{\boldsymbol{\theta}}}.

From the former,
we can derive the distribution of
\eqn{\mathbf{g} \left( \hat{\boldsymbol{\theta}} \right)} as follows:

\deqn{
  	\mathbf{g} \left( \hat{\boldsymbol{\theta}} \right)
  	\approx
  	\mathcal{N}
  	\left(
  	\mathbf{g} \left( \boldsymbol{\theta} \right)
  	,
  	n^{-1}
  	\mathbf{J}
  	\boldsymbol{\Gamma}
  	\mathbf{J}^{\prime}
  	\right)
  }

The uncertainty associated with the estimator
\eqn{\mathbf{g} \left( \hat{\boldsymbol{\theta}} \right)}
is, therefore, given by
\eqn{n^{-1} \mathbf{J} \boldsymbol{\Gamma} \mathbf{J}^{\prime}} .
When \eqn{\boldsymbol{\Gamma}} is unknown,
by substitution,
we can use
the estimated sampling variance-covariance matrix of
\eqn{\hat{\boldsymbol{\theta}}},
that is,
\eqn{\hat{\mathbb{V}} \left( \hat{\boldsymbol{\theta}} \right)}
for \eqn{n^{-1} \boldsymbol{\Gamma}}.
Therefore,
the sampling variance-covariance matrix of
\eqn{\mathbf{g} \left( \hat{\boldsymbol{\theta}} \right)}
is given by

\deqn{
  	\mathbf{g} \left( \hat{\boldsymbol{\theta}} \right)
  	\approx
  	\mathcal{N}
  	\left(
  	\mathbf{g} \left( \boldsymbol{\theta} \right)
  	,
  	\mathbf{J}
  	\hat{\mathbb{V}} \left( \hat{\boldsymbol{\theta}} \right)
  	\mathbf{J}^{\prime}
  	\right) .
  }
}

\subsection{Linear Stochastic Differential Equation Model}{

The measurement model is given by
\deqn{
    \mathbf{y}_{i, t}
    =
    \boldsymbol{\nu}
    +
    \boldsymbol{\Lambda}
    \boldsymbol{\eta}_{i, t}
    +
    \boldsymbol{\varepsilon}_{i, t},
    \quad
    \mathrm{with}
    \quad
    \boldsymbol{\varepsilon}_{i, t}
    \sim
    \mathcal{N}
    \left(
    \mathbf{0},
    \boldsymbol{\Theta}
    \right)
  }
where
\eqn{\mathbf{y}_{i, t}},
\eqn{\boldsymbol{\eta}_{i, t}},
and
\eqn{\boldsymbol{\varepsilon}_{i, t}}
are random variables
and
\eqn{\boldsymbol{\nu}},
\eqn{\boldsymbol{\Lambda}},
and
\eqn{\boldsymbol{\Theta}}
are model parameters.
\eqn{\mathbf{y}_{i, t}}
represents a vector of observed random variables,
\eqn{\boldsymbol{\eta}_{i, t}}
a vector of latent random variables,
and
\eqn{\boldsymbol{\varepsilon}_{i, t}}
a vector of random measurement errors,
at time \eqn{t} and individual \eqn{i}.
\eqn{\boldsymbol{\nu}}
denotes a vector of intercepts,
\eqn{\boldsymbol{\Lambda}}
a matrix of factor loadings,
and
\eqn{\boldsymbol{\Theta}}
the covariance matrix of
\eqn{\boldsymbol{\varepsilon}}.

An alternative representation of the measurement error
is given by
\deqn{
    \boldsymbol{\varepsilon}_{i, t}
    =
    \boldsymbol{\Theta}^{\frac{1}{2}}
    \mathbf{z}_{i, t},
    \quad
    \mathrm{with}
    \quad
    \mathbf{z}_{i, t}
    \sim
    \mathcal{N}
    \left(
    \mathbf{0},
    \mathbf{I}
    \right)
  }
where
\eqn{\mathbf{z}_{i, t}} is a vector of
independent standard normal random variables and
\eqn{
    \left( \boldsymbol{\Theta}^{\frac{1}{2}} \right)
    \left( \boldsymbol{\Theta}^{\frac{1}{2}} \right)^{\prime}
    =
    \boldsymbol{\Theta} .
  }

The dynamic structure is given by
\deqn{
    \mathrm{d} \boldsymbol{\eta}_{i, t}
    =
    \left(
    \boldsymbol{\iota}
    +
    \boldsymbol{\Phi}
    \boldsymbol{\eta}_{i, t}
    \right)
    \mathrm{d}t
    +
    \boldsymbol{\Sigma}^{\frac{1}{2}}
    \mathrm{d}
    \mathbf{W}_{i, t}
  }
where
\eqn{\boldsymbol{\iota}}
is a term which is unobserved and constant over time,
\eqn{\boldsymbol{\Phi}}
is the drift matrix
which represents the rate of change of the solution
in the absence of any random fluctuations,
\eqn{\boldsymbol{\Sigma}}
is the matrix of volatility
or randomness in the process, and
\eqn{\mathrm{d}\boldsymbol{W}}
is a Wiener process or Brownian motion,
which represents random fluctuations.
}
}
\examples{
data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Specific time-interval ----------------------------------------------------
DeltaMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m"
)

# Range of time-intervals ---------------------------------------------------
delta <- DeltaMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m"
)

# Methods -------------------------------------------------------------------
# DeltaMed has a number of methods including
# print, summary, confint, and plot
print(delta)
summary(delta)
confint(delta, level = 0.95)
plot(delta)

}
\references{
Bollen, K. A. (1987).
Total, direct, and indirect effects in structural equation models.
Sociological Methodology, 17, 37.
\doi{10.2307/271028}

Deboeck, P. R., & Preacher, K. J. (2015).
No need to be discrete:
A method for continuous time mediation analysis.
Structural Equation Modeling: A Multidisciplinary Journal, 23 (1), 61–75.
\doi{10.1080/10705511.2014.973960}

Ryan, O., & Hamaker, E. L. (2021).
Time to intervene:
A continuous-time approach to network analysis and centrality.
Psychometrika, 87 (1), 214–252.
\doi{10.1007/s11336-021-09767-0}
}
\seealso{
Other Continuous Time Mediation Functions: 
\code{\link{Direct}()},
\code{\link{Indirect}()},
\code{\link{MCMed}()},
\code{\link{MCPhi}()},
\code{\link{Med}()},
\code{\link{PosteriorMed}()},
\code{\link{TestPhi}()},
\code{\link{TestStable}()},
\code{\link{Total}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Continuous Time Mediation Functions}
\keyword{cTMed}
\keyword{main}
\keyword{uncertainty}
