% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-exp-cov.R
\name{ExpCov}
\alias{ExpCov}
\title{Model-Implied State Covariance Matrix}
\usage{
ExpCov(phi, sigma, delta_t)
}
\arguments{
\item{phi}{Numeric matrix.
The drift matrix (\eqn{\boldsymbol{\Phi}}).
\code{phi} should have row and column names
pertaining to the variables in the system.}

\item{sigma}{Numeric matrix.
The process noise covariance matrix (\eqn{\boldsymbol{\Sigma}}).}

\item{delta_t}{Numeric.
Time interval
(\eqn{\Delta t}).}
}
\value{
Returns a numeric matrix.
}
\description{
The function returns the model-implied state covariance matrix
for a particular time interval \eqn{\Delta t}
given by
\deqn{
  \mathrm{vec} \left( \mathrm{Cov}_{\boldsymbol{\eta}} \right)
  =
  \left(
    \mathbf{J} -
    \boldsymbol{\beta}_{\Delta t} \otimes \boldsymbol{\beta}_{\Delta t}
  \right)^{-1}
  \mathrm{vec} \left( \boldsymbol{\Psi}_{\Delta t} \right)
}
where
\deqn{
  \boldsymbol{\beta}_{\Delta t}
  =
  \exp \left( \Delta t \boldsymbol{\Phi} \right) ,
}
\deqn{
  \boldsymbol{\Psi}_{\Delta t}
  =
  \boldsymbol{\Phi}^{\#}
  \left(
    \exp \left( \Delta t \boldsymbol{\Phi} \right) - \mathbf{J}
  \right)
\mathrm{vec} \left( \boldsymbol{\Sigma} \right) , \quad \mathrm{and}
}
\deqn{
  \boldsymbol{\Phi}^{\#}
  =
  \left(
    \boldsymbol{\Phi} \otimes \mathbf{I}
  \right) +
  \left(
    \mathbf{I} \otimes \boldsymbol{\Phi}
  \right).
}
Note that \eqn{\mathbf{I}} and \eqn{\mathbf{J}} are identity matrices.
}
\examples{
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
sigma <- matrix(
  data = c(
    0.24, 0.02, -0.05,
    0.02, 0.07, 0.02,
    -0.05, 0.02, 0.08
  ),
  nrow = 3
)
delta_t <- 1
ExpCov(
  phi = phi,
  sigma = sigma,
  delta_t = delta_t
)

}
\seealso{
Other Continuous Time Mediation Functions: 
\code{\link{DeltaBeta}()},
\code{\link{DeltaBetaStd}()},
\code{\link{DeltaIndirectCentral}()},
\code{\link{DeltaMed}()},
\code{\link{DeltaMedStd}()},
\code{\link{DeltaTotalCentral}()},
\code{\link{Direct}()},
\code{\link{DirectStd}()},
\code{\link{Indirect}()},
\code{\link{IndirectCentral}()},
\code{\link{IndirectStd}()},
\code{\link{MCBeta}()},
\code{\link{MCBetaStd}()},
\code{\link{MCIndirectCentral}()},
\code{\link{MCMed}()},
\code{\link{MCMedStd}()},
\code{\link{MCPhi}()},
\code{\link{MCTotalCentral}()},
\code{\link{Med}()},
\code{\link{MedStd}()},
\code{\link{PosteriorBeta}()},
\code{\link{PosteriorIndirectCentral}()},
\code{\link{PosteriorMed}()},
\code{\link{PosteriorPhi}()},
\code{\link{PosteriorTotalCentral}()},
\code{\link{Total}()},
\code{\link{TotalCentral}()},
\code{\link{TotalStd}()},
\code{\link{Trajectory}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Continuous Time Mediation Functions}
\keyword{cTMed}
\keyword{expectations}
