% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods.R
\name{plot.ctmeddelta}
\alias{plot.ctmeddelta}
\title{Plot Method for an Object of Class \code{ctmeddelta}}
\usage{
\method{plot}{ctmeddelta}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{Object of class \code{ctmeddelta}.}

\item{alpha}{Numeric.
Significance level}

\item{...}{Additional arguments.}
}
\description{
Plot Method for an Object of Class \code{ctmeddelta}
}
\examples{
data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Range of time-intervals ---------------------------------------------------
delta <- DeltaMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m"
)
plot(delta)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
