% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods.R
\name{plot.ctmedmc}
\alias{plot.ctmedmc}
\title{Plot Method for an Object of Class \code{ctmedmc}}
\usage{
\method{plot}{ctmedmc}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{Object of class \code{ctmedmc}.}

\item{alpha}{Numeric.
Significance level}

\item{...}{Additional arguments.}
}
\description{
Plot Method for an Object of Class \code{ctmedmc}
}
\examples{
data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Range of time-intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
plot(mc)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
