% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods.R
\name{plot.ctmedmed}
\alias{plot.ctmedmed}
\title{Plot Method for an Object of Class \code{ctmedmed}}
\usage{
\method{plot}{ctmedmed}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ctmedmed}.}

\item{...}{Additional arguments.}
}
\description{
Plot Method for an Object of Class \code{ctmedmed}
}
\examples{
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")

# Range of time-intervals ---------------------------------------------------
med <- Med(
  phi = phi,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m"
)
plot(med)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
