% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-posterior-med.R
\name{PosteriorMed}
\alias{PosteriorMed}
\title{Posterior Distribution
of Total, Direct, and Indirect Effects
of X on Y Through M
Over a Specific Time-Interval}
\usage{
PosteriorMed(phi, delta_t, from, to, med, ncores = NULL)
}
\arguments{
\item{phi}{List of numeric matrices.
Each element of the list is a sample
from the posterior distribution
of the drift matrix (\eqn{\boldsymbol{\Phi}}).
Each matrix should have row and column names
pertaining to the variables in the system.}

\item{delta_t}{Numeric.
Time interval
(\eqn{\Delta t}).}

\item{from}{Character string.
Name of the independent variable \eqn{X} in \code{phi}.}

\item{to}{Character string.
Name of the dependent variable \eqn{Y} in \code{phi}.}

\item{med}{Character vector.
Name/s of the mediator variable/s in \code{phi}.}

\item{ncores}{Positive integer.
Number of cores to use.
If \code{ncores = NULL},
use a single core.
Consider using multiple cores
when number of replications \code{R}
is a large value.}
}
\value{
Returns an object
of class \code{ctmedmc} which is a list with the following elements:
\describe{
\item{call}{Function call.}
\item{args}{Function arguments.}
\item{fun}{Function used (PosteriorMed).}
\item{output}{A list with length of \code{length(delta_t)}.}
}
Each element in the \code{output} list has the following elements:
\describe{
\item{est}{Mean of the posterior distribution
of the total, direct, and indirect effects.}
\item{thetahatstar}{Posterior distribution of the
total, direct, and indirect effects.}
}
}
\description{
This function generates a posterior
distribution
of the total, direct and indirect effects
of the independent variable \eqn{X}
on the dependent variable \eqn{Y}
through mediator variables \eqn{\mathbf{m}}
at a particular time-interval \eqn{\Delta t}
using the posterior distribution
of the first-order stochastic differential equation model
drift matrix \eqn{\boldsymbol{\Phi}}.
}
\details{
See \code{\link[=Total]{Total()}},
\code{\link[=Direct]{Direct()}}, and
\code{\link[=Indirect]{Indirect()}} for more details.
}
\examples{
data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$ctsem$posterior_phi

# Specific time-interval ----------------------------------------------------
PosteriorMed(
  phi = phi,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m"
)

# Range of time-intervals ---------------------------------------------------
posterior <- PosteriorMed(
  phi = phi,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m"
)

# Methods -------------------------------------------------------------------
# PosteriorMed has a number of methods including
# print, summary, confint, and plot
print(posterior)
summary(posterior)
confint(posterior, level = 0.95)
plot(posterior)

}
\references{
Bollen, K. A. (1987).
Total, direct, and indirect effects in structural equation models.
Sociological Methodology, 17, 37.
\doi{10.2307/271028}

Deboeck, P. R., & Preacher, K. J. (2015).
No need to be discrete:
A method for continuous time mediation analysis.
Structural Equation Modeling: A Multidisciplinary Journal, 23 (1), 61–75.
\doi{10.1080/10705511.2014.973960}

Ryan, O., & Hamaker, E. L. (2021).
Time to intervene:
A continuous-time approach to network analysis and centrality.
Psychometrika, 87 (1), 214–252.
\doi{10.1007/s11336-021-09767-0}
}
\seealso{
Other Continuous Time Mediation Functions: 
\code{\link{DeltaMed}()},
\code{\link{Direct}()},
\code{\link{Indirect}()},
\code{\link{MCMed}()},
\code{\link{MCPhi}()},
\code{\link{Med}()},
\code{\link{TestPhi}()},
\code{\link{TestStable}()},
\code{\link{Total}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Continuous Time Mediation Functions}
\keyword{cTMed}
\keyword{uncertainty}
