% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-posterior-phi.R
\name{PosteriorPhi}
\alias{PosteriorPhi}
\title{Extract the Posterior Samples of the Drift Matrix}
\usage{
PosteriorPhi(object)
}
\arguments{
\item{object}{Object of class \code{ctStanFit}.
Output of the \code{\link[ctsem:ctStanFit]{ctsem::ctStanFit()}} function.}
}
\value{
Returns an object
of class \code{ctmedposteriorphi} which is a list drift matrices
sampled from the posterior distribution.
}
\description{
The function extracts the posterior samples
of the drift matrix from a fitted model
from the \code{\link[ctsem:ctStanFit]{ctsem::ctStanFit()}} function.
}
\seealso{
Other Continuous Time Mediation Functions: 
\code{\link{DeltaBeta}()},
\code{\link{DeltaBetaStd}()},
\code{\link{DeltaIndirectCentral}()},
\code{\link{DeltaMed}()},
\code{\link{DeltaMedStd}()},
\code{\link{DeltaTotalCentral}()},
\code{\link{Direct}()},
\code{\link{DirectStd}()},
\code{\link{Indirect}()},
\code{\link{IndirectCentral}()},
\code{\link{IndirectStd}()},
\code{\link{MCBeta}()},
\code{\link{MCBetaStd}()},
\code{\link{MCIndirectCentral}()},
\code{\link{MCMed}()},
\code{\link{MCMedStd}()},
\code{\link{MCPhi}()},
\code{\link{MCTotalCentral}()},
\code{\link{Med}()},
\code{\link{MedStd}()},
\code{\link{PosteriorBeta}()},
\code{\link{PosteriorIndirectCentral}()},
\code{\link{PosteriorMed}()},
\code{\link{PosteriorTotalCentral}()},
\code{\link{Total}()},
\code{\link{TotalCentral}()},
\code{\link{TotalStd}()},
\code{\link{Trajectory}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Continuous Time Mediation Functions}
\keyword{cTMed}
\keyword{posterior}
