% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods.R
\name{print.ctmeddelta}
\alias{print.ctmeddelta}
\title{Print Method for Object of Class \code{ctmeddelta}}
\usage{
\method{print}{ctmeddelta}(x, alpha = 0.05, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{ctmeddelta}.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.}

\item{digits}{Integer indicating the number of decimal places to display.}

\item{...}{further arguments.}
}
\value{
Returns a matrix of
time-interval,
estimates,
standard errors,
test statistics,
p-values,
and
confidence intervals.
}
\description{
Print Method for Object of Class \code{ctmeddelta}
}
\examples{
data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Specific time-interval ----------------------------------------------------
delta <- DeltaMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m"
)
print(delta)

# Range of time-intervals ---------------------------------------------------
delta <- DeltaMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m"
)
print(delta)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
