% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods.R
\name{print.ctmedmed}
\alias{print.ctmedmed}
\title{Print Method for Object of Class \code{ctmedmed}}
\usage{
\method{print}{ctmedmed}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{ctmedmed}.}

\item{digits}{Integer indicating the number of decimal places to display.}

\item{...}{further arguments.}
}
\value{
Returns a matrix of effects.
}
\description{
Print Method for Object of Class \code{ctmedmed}
}
\examples{
# ---------------------------------------------------------------------------
# Example 1 -----------------------------------------------------------------
# ---------------------------------------------------------------------------
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")

# Specific time-interval ----------------------------------------------------
med <- Med(
  phi = phi,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m"
)
print(med)

# Range of time-intervals ---------------------------------------------------
med <- Med(
  phi = phi,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m"
)
print(med)

# ---------------------------------------------------------------------------
# Example 2 -----------------------------------------------------------------
# ---------------------------------------------------------------------------
phi <- matrix(
  data = c(
    -6, 5.5, 0, 0,
    1.25, -2.5, 5.9, -7.3,
    0, 0, -6, 2.5,
    5, 0, 0, -6
  ),
  nrow = 4
)
colnames(phi) <- rownames(phi) <- paste0("y", 1:4)

# Specific time-interval ----------------------------------------------------
med <- Med(
  phi = phi,
  delta_t = 1,
  from = "y2",
  to = "y4",
  med = c("y1", "y3")
)
print(med)

# Range of time-intervals ---------------------------------------------------
med <- Med(
  phi = phi,
  delta_t = 1:10,
  from = "y2",
  to = "y4",
  med = c("y1", "y3")
)
print(med)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
