% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods.R
\name{summary.ctmeddelta}
\alias{summary.ctmeddelta}
\title{Summary Method for an Object of Class \code{ctmeddelta}}
\usage{
\method{summary}{ctmeddelta}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{Object of class \code{ctmeddelta}.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of
effects,
time-interval,
estimates,
standard errors,
test statistics,
p-values,
and
confidence intervals.
}
\description{
Summary Method for an Object of Class \code{ctmeddelta}
}
\examples{
data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Specific time-interval ----------------------------------------------------
delta <- DeltaMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m"
)
summary(delta)

# Range of time-intervals ---------------------------------------------------
delta <- DeltaMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m"
)
summary(delta)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
