% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods-ctmedposteriorphi.R
\name{summary.ctmedposteriorphi}
\alias{summary.ctmedposteriorphi}
\title{Summary Method for Object of Class \code{ctmedposteriorphi}}
\usage{
\method{summary}{ctmedposteriorphi}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ctmedposteriorphi}.}

\item{...}{further arguments.}
}
\value{
Returns a list of the posterior means
(in matrix form) and covariance matrix.
}
\description{
Summary Method for Object of Class \code{ctmedposteriorphi}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
