% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods-ctmedtraj.R
\name{summary.ctmedtraj}
\alias{summary.ctmedtraj}
\title{Summary Method for an Object of Class \code{ctmedtraj}}
\usage{
\method{summary}{ctmedtraj}(object, digits = 4, ...)
}
\arguments{
\item{object}{an object of class \code{ctmedtraj}.}

\item{digits}{Integer indicating the number of decimal places to display.}

\item{...}{further arguments.}
}
\value{
Returns a data frame of simulated data.
}
\description{
Summary Method for an Object of Class \code{ctmedtraj}
}
\examples{
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")

traj <- Trajectory(
  mu0 = c(3, 3, -3),
  time = 150,
  phi = phi,
  med = "m"
)

summary(traj)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
