% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestStable}
\alias{TestStable}
\title{Test Stability}
\usage{
TestStable(x)
}
\arguments{
\item{x}{Numeric matrix.}
}
\description{
The function computes the eigenvalues of the input matrix \code{x}.
It checks if the real part of all eigenvalues is negative.
If all eigenvalues have negative real parts,
the system is considered stable.
}
\examples{
x <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
TestStable(x)
x <- matrix(
  data = c(
    -6, 5.5, 0, 0,
    1.25, -2.5, 5.9, -7.3,
    0, 0, -6, 2.5,
    5, 0, 0, -6
  ),
  nrow = 4
)
TestStable(x)

}
\seealso{
Other Continuous Time Mediation Functions: 
\code{\link{DeltaMed}()},
\code{\link{Direct}()},
\code{\link{Indirect}()},
\code{\link{MCMed}()},
\code{\link{MCPhi}()},
\code{\link{Med}()},
\code{\link{PosteriorMed}()},
\code{\link{PosteriorPhi}()},
\code{\link{TestPhi}()},
\code{\link{Total}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Continuous Time Mediation Functions}
\keyword{cTMed}
\keyword{test}
