% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-total-central.R
\name{TotalCentral}
\alias{TotalCentral}
\title{Total Effect Centrality}
\usage{
TotalCentral(phi, delta_t, tol = 0.01)
}
\arguments{
\item{phi}{Numeric matrix.
The drift matrix (\eqn{\boldsymbol{\Phi}}).
\code{phi} should have row and column names
pertaining to the variables in the system.}

\item{delta_t}{Vector of positive numbers.
Time interval
(\eqn{\Delta t}).}

\item{tol}{Numeric.
Smallest possible time interval to allow.}
}
\value{
Returns an object
of class \code{ctmedmed} which is a list with the following elements:
\describe{
\item{call}{Function call.}
\item{args}{Function arguments.}
\item{fun}{Function used ("TotalCentral").}
\item{output}{A matrix of total effect centrality.}
}
}
\description{
Total Effect Centrality
}
\details{
The total effect centrality of a variable
is the sum of the total effects of a variable on all other variables
at a particular time interval.
}
\examples{
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")

# Specific time interval ----------------------------------------------------
TotalCentral(
  phi = phi,
  delta_t = 1
)

# Range of time intervals ---------------------------------------------------
total_central <- TotalCentral(
  phi = phi,
  delta_t = 1:30
)
plot(total_central)

# Methods -------------------------------------------------------------------
# TotalCentral has a number of methods including
# print, summary, and plot
total_central <- TotalCentral(
  phi = phi,
  delta_t = 1:5
)
print(total_central)
summary(total_central)
plot(total_central)

}
\references{
Bollen, K. A. (1987).
Total, direct, and indirect effects in structural equation models.
Sociological Methodology, 17, 37.
\doi{10.2307/271028}

Deboeck, P. R., & Preacher, K. J. (2015).
No need to be discrete:
A method for continuous time mediation analysis.
Structural Equation Modeling: A Multidisciplinary Journal, 23 (1), 61-75.
\doi{10.1080/10705511.2014.973960}

Pesigan, I. J. A., Russell, M. A., & Chow, S.-M. (2025).
Inferences and effect sizes for direct, indirect, and total effects
in continuous-time mediation models.
Psychological Methods.
\doi{10.1037/met0000779}

Ryan, O., & Hamaker, E. L. (2021).
Time to intervene:
A continuous-time approach to network analysis and centrality.
Psychometrika, 87 (1), 214-252.
\doi{10.1007/s11336-021-09767-0}
}
\seealso{
Other Continuous-Time Mediation Functions:
\code{\link[=BootBeta]{BootBeta()}},
\code{\link[=BootBetaStd]{BootBetaStd()}},
\code{\link[=BootDirectCentral]{BootDirectCentral()}},
\code{\link[=BootIndirectCentral]{BootIndirectCentral()}},
\code{\link[=BootMed]{BootMed()}},
\code{\link[=BootMedStd]{BootMedStd()}},
\code{\link[=BootTotalCentral]{BootTotalCentral()}},
\code{\link[=DeltaBeta]{DeltaBeta()}},
\code{\link[=DeltaBetaStd]{DeltaBetaStd()}},
\code{\link[=DeltaDirectCentral]{DeltaDirectCentral()}},
\code{\link[=DeltaIndirectCentral]{DeltaIndirectCentral()}},
\code{\link[=DeltaMed]{DeltaMed()}},
\code{\link[=DeltaMedStd]{DeltaMedStd()}},
\code{\link[=DeltaTotalCentral]{DeltaTotalCentral()}},
\code{\link[=Direct]{Direct()}},
\code{\link[=DirectCentral]{DirectCentral()}},
\code{\link[=DirectStd]{DirectStd()}},
\code{\link[=Indirect]{Indirect()}},
\code{\link[=IndirectCentral]{IndirectCentral()}},
\code{\link[=IndirectStd]{IndirectStd()}},
\code{\link[=MCBeta]{MCBeta()}},
\code{\link[=MCBetaStd]{MCBetaStd()}},
\code{\link[=MCDirectCentral]{MCDirectCentral()}},
\code{\link[=MCIndirectCentral]{MCIndirectCentral()}},
\code{\link[=MCMed]{MCMed()}},
\code{\link[=MCMedStd]{MCMedStd()}},
\code{\link[=MCPhi]{MCPhi()}},
\code{\link[=MCPhiSigma]{MCPhiSigma()}},
\code{\link[=MCTotalCentral]{MCTotalCentral()}},
\code{\link[=Med]{Med()}},
\code{\link[=MedStd]{MedStd()}},
\code{\link[=PosteriorBeta]{PosteriorBeta()}},
\code{\link[=PosteriorDirectCentral]{PosteriorDirectCentral()}},
\code{\link[=PosteriorIndirectCentral]{PosteriorIndirectCentral()}},
\code{\link[=PosteriorMed]{PosteriorMed()}},
\code{\link[=PosteriorTotalCentral]{PosteriorTotalCentral()}},
\code{\link[=Total]{Total()}},
\code{\link[=TotalStd]{TotalStd()}},
\code{\link[=Trajectory]{Trajectory()}}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Continuous-Time Mediation Functions}
\keyword{cTMed}
\keyword{effects}
\keyword{network}
