% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-total-central-std.R
\name{TotalCentralStd}
\alias{TotalCentralStd}
\title{Standardized Total Effect Centrality}
\usage{
TotalCentralStd(phi, sigma, delta_t, tol = 0.001)
}
\arguments{
\item{phi}{Numeric matrix.
The drift matrix (\eqn{\boldsymbol{\Phi}}).
\code{phi} should have row and column names
pertaining to the variables in the system.}

\item{sigma}{Numeric matrix.
The process noise covariance matrix (\eqn{\boldsymbol{\Sigma}}).}

\item{delta_t}{Vector of positive numbers.
Time interval
(\eqn{\Delta t}).}

\item{tol}{Numeric.
Smallest possible time interval to allow.}
}
\value{
Returns an object
of class \code{ctmedmed} which is a list with the following elements:
\describe{
\item{call}{Function call.}
\item{args}{Function arguments.}
\item{fun}{Function used ("TotalCentralStd").}
\item{output}{A matrix of standardized total effect centrality.}
}
}
\description{
Standardized Total Effect Centrality
}
\details{
The standardized total effect centrality of a variable
is the sum of the standardized total effects of a variable
on all other variables at a particular time interval.
}
\examples{
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
sigma <- matrix(
  data = c(
    0.24455556, 0.02201587, -0.05004762,
    0.02201587, 0.07067800, 0.01539456,
    -0.05004762, 0.01539456, 0.07553061
  ),
  nrow = 3
)

# Specific time interval ----------------------------------------------------
TotalCentralStd(
  phi = phi,
  sigma = sigma,
  delta_t = 1
)

# Range of time intervals ---------------------------------------------------
total_central_std <- TotalCentralStd(
  phi = phi,
  sigma = sigma,
  delta_t = 1:30
)
plot(total_central_std)

# Methods -------------------------------------------------------------------
# TotalCentralStd has a number of methods including
# print, summary, and plot
total_central_std <- TotalCentralStd(
  phi = phi,
  sigma = sigma,
  delta_t = 1:5
)
print(total_central_std)
summary(total_central_std)
plot(total_central_std)

}
\seealso{
Other Continuous-Time Mediation Functions:
\code{\link[=BootBeta]{BootBeta()}},
\code{\link[=BootBetaStd]{BootBetaStd()}},
\code{\link[=BootDirectCentral]{BootDirectCentral()}},
\code{\link[=BootIndirectCentral]{BootIndirectCentral()}},
\code{\link[=BootMed]{BootMed()}},
\code{\link[=BootMedStd]{BootMedStd()}},
\code{\link[=BootTotalCentral]{BootTotalCentral()}},
\code{\link[=DeltaBeta]{DeltaBeta()}},
\code{\link[=DeltaBetaStd]{DeltaBetaStd()}},
\code{\link[=DeltaDirectCentral]{DeltaDirectCentral()}},
\code{\link[=DeltaIndirectCentral]{DeltaIndirectCentral()}},
\code{\link[=DeltaMed]{DeltaMed()}},
\code{\link[=DeltaMedStd]{DeltaMedStd()}},
\code{\link[=DeltaTotalCentral]{DeltaTotalCentral()}},
\code{\link[=Direct]{Direct()}},
\code{\link[=DirectCentral]{DirectCentral()}},
\code{\link[=DirectCentralStd]{DirectCentralStd()}},
\code{\link[=DirectStd]{DirectStd()}},
\code{\link[=Indirect]{Indirect()}},
\code{\link[=IndirectCentral]{IndirectCentral()}},
\code{\link[=IndirectCentralStd]{IndirectCentralStd()}},
\code{\link[=IndirectStd]{IndirectStd()}},
\code{\link[=MCBeta]{MCBeta()}},
\code{\link[=MCBetaStd]{MCBetaStd()}},
\code{\link[=MCDirectCentral]{MCDirectCentral()}},
\code{\link[=MCIndirectCentral]{MCIndirectCentral()}},
\code{\link[=MCMed]{MCMed()}},
\code{\link[=MCMedStd]{MCMedStd()}},
\code{\link[=MCPhi]{MCPhi()}},
\code{\link[=MCPhiSigma]{MCPhiSigma()}},
\code{\link[=MCTotalCentral]{MCTotalCentral()}},
\code{\link[=Med]{Med()}},
\code{\link[=MedStd]{MedStd()}},
\code{\link[=PosteriorBeta]{PosteriorBeta()}},
\code{\link[=PosteriorDirectCentral]{PosteriorDirectCentral()}},
\code{\link[=PosteriorIndirectCentral]{PosteriorIndirectCentral()}},
\code{\link[=PosteriorMed]{PosteriorMed()}},
\code{\link[=PosteriorTotalCentral]{PosteriorTotalCentral()}},
\code{\link[=Total]{Total()}},
\code{\link[=TotalCentral]{TotalCentral()}},
\code{\link[=TotalStd]{TotalStd()}},
\code{\link[=Trajectory]{Trajectory()}}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Continuous-Time Mediation Functions}
\keyword{cTMed}
\keyword{effects}
\keyword{network}
