% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-trajectory.R
\name{Trajectory}
\alias{Trajectory}
\title{Simulate Trajectories of Variables}
\usage{
Trajectory(mu0, time, phi, med)
}
\arguments{
\item{mu0}{Numeric vector.
Initial values of the variables.}

\item{time}{Positive integer.
Number of time points.}

\item{phi}{Numeric matrix.
The drift matrix (\eqn{\boldsymbol{\Phi}}).
\code{phi} should have row and column names
pertaining to the variables in the system.}

\item{med}{Character vector.
Name/s of the mediator variable/s in \code{phi}.}
}
\value{
Returns an object
of class \code{ctmedtraj} which is a list with the following elements:
\describe{
\item{call}{Function call.}
\item{args}{Function arguments.}
\item{fun}{Function used ("Trajectory").}
\item{output}{A data frame of simulated data.}
}
}
\description{
This function simulates trajectories of variables
without measurement error or process noise.
\code{Total} corresponds to the total effect
and \code{Direct} corresponds to the portion of the total effect
where the indirect effect is removed.
}
\examples{
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")

traj <- Trajectory(
  mu0 = c(3, 3, -3),
  time = 150,
  phi = phi,
  med = "m"
)
plot(traj)

# Methods -------------------------------------------------------------------
# Trajectory has a number of methods including
# print, summary, and plot

traj <- Trajectory(
  mu0 = c(3, 3, -3),
  time = 25,
  phi = phi,
  med = "m"
)
print(traj)
summary(traj)
plot(traj)

}
\seealso{
Other Continuous-Time Mediation Functions:
\code{\link[=BootBeta]{BootBeta()}},
\code{\link[=BootBetaStd]{BootBetaStd()}},
\code{\link[=BootDirectCentral]{BootDirectCentral()}},
\code{\link[=BootIndirectCentral]{BootIndirectCentral()}},
\code{\link[=BootMed]{BootMed()}},
\code{\link[=BootMedStd]{BootMedStd()}},
\code{\link[=BootTotalCentral]{BootTotalCentral()}},
\code{\link[=DeltaBeta]{DeltaBeta()}},
\code{\link[=DeltaBetaStd]{DeltaBetaStd()}},
\code{\link[=DeltaDirectCentral]{DeltaDirectCentral()}},
\code{\link[=DeltaIndirectCentral]{DeltaIndirectCentral()}},
\code{\link[=DeltaMed]{DeltaMed()}},
\code{\link[=DeltaMedStd]{DeltaMedStd()}},
\code{\link[=DeltaTotalCentral]{DeltaTotalCentral()}},
\code{\link[=Direct]{Direct()}},
\code{\link[=DirectCentral]{DirectCentral()}},
\code{\link[=DirectStd]{DirectStd()}},
\code{\link[=Indirect]{Indirect()}},
\code{\link[=IndirectCentral]{IndirectCentral()}},
\code{\link[=IndirectStd]{IndirectStd()}},
\code{\link[=MCBeta]{MCBeta()}},
\code{\link[=MCBetaStd]{MCBetaStd()}},
\code{\link[=MCDirectCentral]{MCDirectCentral()}},
\code{\link[=MCIndirectCentral]{MCIndirectCentral()}},
\code{\link[=MCMed]{MCMed()}},
\code{\link[=MCMedStd]{MCMedStd()}},
\code{\link[=MCPhi]{MCPhi()}},
\code{\link[=MCPhiSigma]{MCPhiSigma()}},
\code{\link[=MCTotalCentral]{MCTotalCentral()}},
\code{\link[=Med]{Med()}},
\code{\link[=MedStd]{MedStd()}},
\code{\link[=PosteriorBeta]{PosteriorBeta()}},
\code{\link[=PosteriorDirectCentral]{PosteriorDirectCentral()}},
\code{\link[=PosteriorIndirectCentral]{PosteriorIndirectCentral()}},
\code{\link[=PosteriorMed]{PosteriorMed()}},
\code{\link[=PosteriorTotalCentral]{PosteriorTotalCentral()}},
\code{\link[=Total]{Total()}},
\code{\link[=TotalCentral]{TotalCentral()}},
\code{\link[=TotalStd]{TotalStd()}}
}
\concept{Continuous-Time Mediation Functions}
\keyword{cTMed}
\keyword{effects}
\keyword{path}
