% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-package.R
\docType{package}
\name{cTMed-package}
\alias{cTMed}
\alias{cTMed-package}
\title{cTMed: Continuous-Time Mediation}
\description{
Computes effect sizes, standard errors, and confidence intervals for total, direct, and indirect effects in continuous-time mediation models as described in Pesigan, Russell, and Chow (2025) \doi{10.1037/met0000779}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/cTMed}
  \item \url{https://jeksterslab.github.io/cTMed/}
  \item Report bugs at \url{https://github.com/jeksterslab/cTMed/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Michael A. Russell \email{mar60@psu.edu} (\href{https://orcid.org/0000-0002-3956-604X}{ORCID}) [contributor]
  \item Sy-Miin Chow \email{quc16@psu.edu} (\href{https://orcid.org/0000-0003-1938-027X}{ORCID}) [contributor]
}

}
\keyword{internal}
