% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-package.R
\docType{package}
\name{cTMed-package}
\alias{cTMed}
\alias{cTMed-package}
\title{cTMed: Continuous Time Mediation}
\description{
Calculates standard errors and confidence intervals for effects in continuous time mediation models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/cTMed}
  \item \url{https://jeksterslab.github.io/cTMed/}
  \item Report bugs at \url{https://github.com/jeksterslab/cTMed/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
