% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods.R
\name{plot.ctmedmed}
\alias{plot.ctmedmed}
\title{Plot Method for an Object of Class \code{ctmedmed}}
\usage{
\method{plot}{ctmedmed}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ctmedmed}.}

\item{...}{Additional arguments.}
}
\description{
Plot Method for an Object of Class \code{ctmedmed}
}
\examples{
# ---------------------------------------------------------------------------
# Example 1 -----------------------------------------------------------------
# ---------------------------------------------------------------------------
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")

# Range of time-intervals ---------------------------------------------------
med <- Med(
  phi = phi,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m"
)
plot(med)

# ---------------------------------------------------------------------------
# Example 2 -----------------------------------------------------------------
# ---------------------------------------------------------------------------
phi <- matrix(
  data = c(
    -6, 5.5, 0, 0,
    1.25, -2.5, 5.9, -7.3,
    0, 0, -6, 2.5,
    5, 0, 0, -6
  ),
  nrow = 4
)
colnames(phi) <- rownames(phi) <- paste0("y", 1:4)

# Range of time-intervals ---------------------------------------------------
med <- Med(
  phi = phi,
  delta_t = seq(from = 0, to = 5, length.out = 500),
  from = "y2",
  to = "y4",
  med = c("y1", "y3")
)
plot(med)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
