% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-methods.R
\name{plot.ctmedmed}
\alias{plot.ctmedmed}
\title{Plot Method for an Object of Class \code{ctmedmed}}
\usage{
\method{plot}{ctmedmed}(
  x,
  col_direct = "#2c7bb6",
  col_indirect = "#d7191c",
  col_total = "#5e3c99",
  ...
)
}
\arguments{
\item{x}{Object of class \code{ctmedmed}.}

\item{col_direct}{Character string.
Optional argument.
Color for the direct effect.}

\item{col_indirect}{Character string.
Optional argument.
Color for the indirect effect.}

\item{col_total}{Character string.
Optional argument.
Color for the total effect.}

\item{...}{Additional arguments.}
}
\description{
Plot Method for an Object of Class \code{ctmedmed}
}
\examples{
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")

# Range of time-intervals ---------------------------------------------------
med <- Med(
  phi = phi,
  delta_t = 1:30,
  from = "x",
  to = "y",
  med = "m"
)
plot(med)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
