% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-posterior-phi.R
\name{PosteriorPhi}
\alias{PosteriorPhi}
\title{Extract the Posterior Samples of the Drift Matrix}
\usage{
PosteriorPhi(x)
}
\arguments{
\item{x}{Object of class \code{ctStanFit}.
Output of the \code{\link[ctsem:ctStanFit]{ctsem::ctStanFit()}} function.}
}
\description{
The function extracts the posterior samples
of the drift matrix from a fitted model
using the \code{\link[ctsem:ctStanFit]{ctsem::ctStanFit()}} function.
}
\seealso{
Other Continuous Time Mediation Functions: 
\code{\link{DeltaMed}()},
\code{\link{Direct}()},
\code{\link{Indirect}()},
\code{\link{MCMed}()},
\code{\link{MCPhi}()},
\code{\link{Med}()},
\code{\link{PosteriorMed}()},
\code{\link{TestPhi}()},
\code{\link{TestStable}()},
\code{\link{Total}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Continuous Time Mediation Functions}
\keyword{cTMed}
\keyword{uncertainty}
