% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTMed-posterior-phi.R
\name{PosteriorPhi}
\alias{PosteriorPhi}
\title{Extract the Posterior Samples of the Drift Matrix}
\usage{
PosteriorPhi(object)
}
\arguments{
\item{object}{Object of class \code{ctStanFit}.
Output of the \code{\link[ctsem:ctStanFit]{ctsem::ctStanFit()}} function.}
}
\description{
The function extracts the posterior samples
of the drift matrix from a fitted model
from the \code{\link[ctsem:ctStanFit]{ctsem::ctStanFit()}} function.
}
\seealso{
Other Continuous Time Mediation Functions: 
\code{\link{DeltaIndirectCentral}()},
\code{\link{DeltaMed}()},
\code{\link{DeltaTotalCentral}()},
\code{\link{Direct}()},
\code{\link{Indirect}()},
\code{\link{IndirectCentral}()},
\code{\link{MCIndirectCentral}()},
\code{\link{MCMed}()},
\code{\link{MCPhi}()},
\code{\link{MCTotalCentral}()},
\code{\link{Med}()},
\code{\link{PosteriorIndirectCentral}()},
\code{\link{PosteriorMed}()},
\code{\link{PosteriorTotalCentral}()},
\code{\link{TestPhi}()},
\code{\link{TestStable}()},
\code{\link{Total}()},
\code{\link{TotalCentral}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Continuous Time Mediation Functions}
\keyword{cTMed}
\keyword{uncertainty}
